saving Passed maf

+++ source/variants/filtering.py
@@ -582,32 +582,49 @@ def filter_for_variant_caller(caller, cnf, bcbio_structure):
         # sample.filtered_tsv_by_callername[caller.name] = tsv
         # sample.filtered_maf_by_callername[caller.name] = maf
 
-    comb_maf_fpath = join(bcbio_structure.var_dirpath, caller.name + '.maf')
-    caller.combined_filt_maf_fpath = combine_mafs(cnf, caller.find_filt_maf_by_sample().values(), comb_maf_fpath)
-    comb_maf_fpath_symlink = join(bcbio_structure.date_dirpath, caller.name + '.maf')
+    caller.combined_filt_maf_fpath, caller.combined_filt_pass_maf_fpath = \
+        combine_mafs(cnf,
+                     caller.find_filt_maf_by_sample().values(),
+                     join(bcbio_structure.var_dirpath, caller.name))
+
+    comb_basefname = basename(caller.combined_filt_maf_fpath)
+    pass_comb_basefname = basename(caller.combined_filt_pass_maf_fpath)
+
+    comb_maf_fpath_symlink = join(bcbio_structure.date_dirpath, comb_basefname)
+    comb_pass_maf_fpath_symlink = join(bcbio_structure.date_dirpath, pass_comb_basefname)
+
     if not exists(comb_maf_fpath_symlink):
-        os.symlink(comb_maf_fpath, comb_maf_fpath_symlink)
+        os.symlink(caller.combined_filt_maf_fpath, comb_maf_fpath_symlink)
+    if not exists(comb_pass_maf_fpath_symlink):
+        os.symlink(caller.combined_filt_pass_maf_fpath, comb_pass_maf_fpath_symlink)
+
     info('-' * 70)
     info()
 
     return caller
 
 
-def combine_mafs(cnf, maf_fpaths, output_fpath):
-    if isfile(output_fpath):
-        os.remove(output_fpath)
+def combine_mafs(cnf, maf_fpaths, output_basename):
+    output_fpath = output_basename + '.maf'
+    output_pass_fpath = output_basename + '.pass.maf'
 
-    if not isdir(dirname(output_fpath)):
-        safe_mkdir(dirname(output_fpath))
+    if isfile(output_fpath): os.remove(output_fpath)
+    if isfile(output_pass_fpath): os.remove(output_pass_fpath)
 
-    with open(output_fpath, 'w') as out:
+    if not isdir(dirname(output_fpath)): safe_mkdir(dirname(output_fpath))
+
+    with open(output_fpath, 'w') as out, \
+         open(output_pass_fpath, 'w') as out_pass:
+
         for i, fpath in enumerate(maf_fpaths):
             with open(fpath) as inp:
                 for j, line in enumerate(inp):
                     if i > 0 and j in [0, 1]:
                         continue
                     out.write(line)
-    return output_fpath
+                    if '\tInvalid\t' not in line:
+                        out_pass.write(line)
+    return output_fpath, output_pass_fpath
 
 
 def postprocess_vcf(sample, original_anno_vcf_fpath, work_filt_vcf_fpath):
@@ -627,11 +644,11 @@ def postprocess_vcf(sample, original_anno_vcf_fpath, work_filt_vcf_fpath):
     final_vcf_fpath = add_suffix(original_anno_vcf_fpath, 'filt').replace('varAnnotate', 'varFilter')
     safe_mkdir(dirname(final_vcf_fpath))
 
-    file_basepath = splitext(final_vcf_fpath)[0]
-    final_vcf_fpath = file_basepath + '.vcf'
-    final_tsv_fpath = file_basepath + '.tsv'
-    final_maf_fpath = file_basepath + '.maf'
-    final_clean_vcf_fpath = file_basepath + '.passed.vcf'  # for futrher processing
+    file_basefpath = splitext(final_vcf_fpath)[0]
+    final_vcf_fpath = file_basefpath + '.vcf'
+    final_tsv_fpath = file_basefpath + '.tsv'
+    final_maf_fpath = file_basefpath + '.maf'
+    final_pass_vcf_fpath = join(dirname(final_vcf_fpath), sample.name + BCBioStructure.pass_filt_vcf_ending)  # for futrher processing
 
     BCBioStructure.move_vcfs_to_var(sample)
 
@@ -643,14 +660,14 @@ def postprocess_vcf(sample, original_anno_vcf_fpath, work_filt_vcf_fpath):
     igvtools_index(cnf, final_vcf_fpath)
 
     # Cleaning rejected variants
-    clean_filtered_vcf_fpath = remove_rejected(cnf, work_filt_vcf_fpath)
-    if vcf_is_empty(cnf, clean_filtered_vcf_fpath):
+    pass_filtered_vcf_fpath = remove_rejected(cnf, work_filt_vcf_fpath)
+    if vcf_is_empty(cnf, pass_filtered_vcf_fpath):
         info('All variants are rejected.')
-    if isfile(final_clean_vcf_fpath): os.remove(final_clean_vcf_fpath)
-    if islink(clean_filtered_vcf_fpath): os.unlink(clean_filtered_vcf_fpath)
-    shutil.copy(clean_filtered_vcf_fpath, final_clean_vcf_fpath)
+    if isfile(final_pass_vcf_fpath): os.remove(final_pass_vcf_fpath)
+    if islink(pass_filtered_vcf_fpath): os.unlink(pass_filtered_vcf_fpath)
+    shutil.copy(pass_filtered_vcf_fpath, final_pass_vcf_fpath)
     # os.symlink(final_clean_vcf_fpath, clean_filtered_vcf_fpath)
-    igvtools_index(cnf, final_clean_vcf_fpath)
+    igvtools_index(cnf, final_pass_vcf_fpath)
 
     # Converting to TSV
     if work_filt_vcf_fpath and 'tsv_fields' in cnf:
